requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
appium-python-client==2.6.0
async-generator==1.10; python_version >= '3.5'
atomicwrites==1.4.1; sys_platform == 'win32'
attrs==21.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
black==22.6.0
certifi==2022.6.15
cffi==1.15.1; os_name == 'nt' and implementation_name != 'pypy'
click==8.1.3; python_full_version >= '3.7.0'
colorama==0.4.5; sys_platform == 'win32'
cryptography==37.0.4
h11==0.13.0; python_version >= '3.6'
idna==3.3; python_version >= '3.5'
iniconfig==1.1.1
mypy-extensions==0.4.3
outcome==1.2.0; python_full_version >= '3.7.0'
packaging==21.3; python_version >= '3.6'
pathspec==0.9.0
platformdirs==2.5.2; python_full_version >= '3.7.0'
pluggy==1.0.0; python_version >= '3.6'
py==1.11.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pycparser==2.21
pyopenssl==22.0.0
pyparsing==3.0.9; python_full_version >= '3.6.8'
pysocks==1.7.1
pytest==7.1.2
selenium==4.3.0; python_version ~= '3.7'
sniffio==1.2.0; python_version >= '3.5'
sortedcontainers==2.4.0
tomli==2.0.1; python_full_version >= '3.7.0'
trio-websocket==0.9.2; python_version >= '3.5'
trio==0.21.0; python_full_version >= '3.7.0'
urllib3[secure,socks]==1.26.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
wsproto==1.1.0; python_full_version >= '3.7.0'
